requirements.txt
annotated-types==0.7.0
anyio==4.13.0
bitarray==3.8.1
certifi==2026.2.25
charset-normalizer==3.4.7
ckzg==2.1.7
cytoolz==1.1.0
eth-account==0.13.7
eth-hash==0.8.0
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==6.0.0
eth-utils==6.0.0
eth_abi==6.0.0b1
h11==0.16.0
h2==4.3.0
hexbytes==1.3.1
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
packaging==26.0
parsimonious==0.10.0
pluggy==1.6.0
poly_eip712_structs==0.0.1
py_builder_signing_sdk==0.0.2
py_clob_client==0.34.6
py_order_utils==0.3.2
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pytest==9.0.3
python-dotenv==1.2.2
regex==2026.4.4
requests==2.33.1
rlp==4.1.0
toolz==1.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
psycopg2-binary==2.9.10
